adding device tests
